+++ babel.config.js
@@ -1,6 +1,10 @@
 module.exports = {
-  presets: [
-    ['@babel/preset-env', {targets: {node: 'current'}}],
-     '@babel/preset-typescript',
-  ],
+    presets: [
+        ['@babel/preset-env', { targets: { node: 'current' } }],
+        '@babel/preset-typescript',
+        '@babel/preset-react'
+    ],
+    plugins: [
+        'angularjs-annotate'
+    ]
 };