+++ src/datasources/bpftrace/script.ts
@@ -5,7 +5,8 @@ import { TargetFormat } from '../lib/models/datasource';
 
 enum MetricType {
     Histogram = "histogram",
-    Output = "output"
+    Output = "output",
+    Stacks = "stacks"
 }
 
 export enum ScriptStatus {
@@ -25,7 +26,6 @@ export default class BPFtraceScript {
 
     // additional properties by ScriptRegistry
     readonly code: string;
-    lastRequested: number;
 
     constructor(registerResponse: any, code: string) {
         this.name = registerResponse.name;
@@ -35,7 +35,6 @@ export default class BPFtraceScript {
         this.output = registerResponse.output;
 
         this.code = code;
-        this.lastRequested = new Date().getTime();
     }
 
     hasFailed() {
@@ -66,9 +65,13 @@ export default class BPFtraceScript {
         return null;
     }
 
+    getAllDataMetrics() {
+        return this.vars.map(var_ => this.getDataMetric(var_));
+    }
+
     async getDataMetrics(pmapiSrv: PmapiSrv, format: TargetFormat) {
         if (format === TargetFormat.TimeSeries) {
-            return this.vars.map(var_ => this.getDataMetric(var_));
+            return this.getAllDataMetrics();
         }
         else if (format === TargetFormat.Heatmap) {
             const metric = await this.findMetricForMetricType(pmapiSrv, MetricType.Histogram);
@@ -80,6 +83,12 @@ export default class BPFtraceScript {
             const metric = await this.findMetricForMetricType(pmapiSrv, MetricType.Output);
             if (!metric)
                 throw new Error("Please printf() a table in CSV format in the BPFtrace script.");
+            return [metric];
+        }
+        else if (format === TargetFormat.FlameGraph) {
+            const metric = await this.findMetricForMetricType(pmapiSrv, MetricType.Stacks);
+            if (!metric)
+                throw new Error("Cannot find any sampled stacks in this BPFtrace script. Try: profile:hz:99 { @[kstack] = count(); }");
             return [metric];
         }
         throw new Error("Unsupported panel format.");