+++ src/datasources/lib/services/poll_srv.ts
@@ -3,13 +3,14 @@ import DataStore from "../datastore";
 import { PmapiSrv, MissingMetricsError } from "./pmapi_srv";
 
 export default class PollSrv {
-    private requestedMetrics: Record<string, number> = {}; // {metric: lastRequested}
+    private requestedMetrics: string[] = [];
 
-    constructor(private pmapiSrv: PmapiSrv, private datastore: DataStore, private keepPollingMs: number) {
+    constructor(private pmapiSrv: PmapiSrv, private datastore: DataStore) {
     }
 
     async poll() {
-        const metrics = Object.keys(this.requestedMetrics);
+        // copy metrics array, could change between API calls
+        const metrics = this.requestedMetrics;
         if (metrics.length === 0) {
             return;
         }
@@ -21,35 +22,22 @@ export default class PollSrv {
         const missingMetrics = _.difference(metrics, returnedMetrics);
         if (missingMetrics.length > 0) {
             console.debug(`fetch didn't include result for ${missingMetrics.join(', ')}, clearing it from requested metrics`);
-            for (const missingMetric of missingMetrics) {
-                delete this.requestedMetrics[missingMetric];
-            }
+            this.requestedMetrics = _.difference(this.requestedMetrics, missingMetrics);
         }
     }
 
     async ensurePolling(metrics: string[]) {
         const metadatas = await this.pmapiSrv.getMetricMetadatas(metrics);
-        const missingMetrics = _.difference(metrics, Object.keys(metadatas));
+        const validMetrics = Object.keys(metadatas);
+        const missingMetrics = _.difference(metrics, validMetrics);
         if (missingMetrics.length > 0) {
             throw new MissingMetricsError(missingMetrics);
         }
-
-        const validMetrics = _.intersection(metrics, Object.keys(metadatas));
-        const now = new Date().getTime();
-        for (const metric of validMetrics) {
-            this.requestedMetrics[metric] = now;
-        }
+        this.requestedMetrics = _.union(this.requestedMetrics, validMetrics);
     }
 
     removeMetricsFromPolling(metrics: string[]) {
-        for (const metric of metrics) {
-            if (metric in this.requestedMetrics)
-                delete this.requestedMetrics[metric];
-        }
+        this.requestedMetrics = _.difference(this.requestedMetrics, metrics);
     }
 
-    cleanup() {
-        const pollExpiry = new Date().getTime() - this.keepPollingMs;
-        this.requestedMetrics = _.pickBy(this.requestedMetrics, (lastRequested: number) => lastRequested > pollExpiry);
-    }
 }